AgenticAI/adventure-game/requirements.txt
fastmcp
pyswip
httpx
python-dotenv
strands-agents[ollama]
fastapi
uvicorn[standard]
